Fixed major freeze bug
When changing layer to an uncached layer, the program would freeze for up to 4 seconds.  This was caused by the gitPixel function in EMimage.draw being called twice.  This line was retreiving the color of an unloaded pixel, but only pos was being used, not color.  Added new function getPos for when you only want the x and y not the pixel